Deserialize active user as the correct custom user type
When using the asynchronous login and logout APIs, the active user is
correctly mapped using the configured custom `User` subclass. However on
initialisation, when the active user is deserialised from user defaults,
the custom subclass is ignored. This can result in surprising bugs or
crashes when callers expect their custom `User` subclassed to be used
on app launch.

Because `Client.userType` is specified dynamically, rather than
statically, it's not possible to parameterise the `Mapper` using the
dynamic type (dynamic types can't be used in static declarations). Thus
in order to configure a mapper of dynamic type, we need to use type
erasure: this is achieved using the `AnyMapper` type.

Swift properties can't have generic type parameters in the same way that
functions can, so in order to parameterise the setter a new generic
function `setUserType(_:)` has been introduced. This internally
constructs the type-erased `userMapper`, which can later be accessed
when deserialising from user defaults, and the mapped object dynamically
cast back to the expected `User` subclass.