Realm and RealmSwift Update Issues Work Around
Realm and RealmSwift have both been updated from 2.4.2, to 2.4.3 few weeks AFTER this Kinvey iOS SDK version 3.3.8 was released. Now when a new project has pods installed or a existing project updates its Pods, the version of Realm and RealmSwift as updated, after which point. Calls to Collection.find() fail. This is due to the update of Realm/RealmSwift, when explicitly stating 2.4.2 as the Realm and RealmSwift version with the latest (3.3.8) SDK, explicitly stating the versions for both these dependencies 
is the only way to make it work, but as this is "extra work" for the users I thought to add this work around here, it may buy the Kinvey Engineers to figure out if the problem with the Realm update is a Kinvey error or a Realm error.
Note: this a Work around to stop all iOS users becoming affected.